storage/requirements.txt
requests==2.26.0
connexion==2.9.0
swagger-ui-bundle==0.0.9
SQLAlchemy==1.3.22
mysql-connector-python==8.0.23
pymysql==1.0.2